A little bit of refactoring - separated iodined.c into separate files Fixed up client side compile issues Removed old packet handling code - TODO: use sliding window buffer instead

+++ src/encoding.h
@@ -35,7 +35,7 @@ struct encoder {
 	int (*blocksize_encoded)(void);
 };
 
-int build_hostname(char *, size_t, const char *, const size_t, const char *, struct encoder *, int);
+int build_hostname(char *, size_t, const char *, const size_t, const char *, struct encoder *, size_t);
 int unpack_data(char *, size_t, char *, size_t, struct encoder *);
 int inline_dotify(char *, size_t);
 int inline_undotify(char *, size_t);
